Racket*:several updates
I add a `Racket_concurrency` version, which is much faster than the `Racket_parallel` version.
Additionally, a few updates and fixes are applied to the `Racket` and `Racket_parallel` version.
Except for all the optimizations, fixing the position of the `--no-compiled` flag in the `Racket` version leads to
an extraodinarily poor performance due to the absence of compiled code.